Rename project to Video Transcoder
The project will primary be focused on encoding video to
other formats, providing options such as codec, video quality
or size, output format, cutting video, etc.
