+++ README.md
@@ -1,4 +1,4 @@
-# Video Editor
+# Video Transcoder
 [![Build Status](https://travis-ci.org/brarcher/VideoEditor.svg?branch=master)](https://travis-ci.org/brarcher/VideoEditor)
 
 # Thanks
