+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources
     xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
 
-    <string name="app_name">Video Editor</string>
+    <string name="app_name">Video Transcoder</string>
     <string name="rangeSeekBarText">Use RangeSeekbar to select time range for cutting video or extracting images from video(rangeseekbar ineffective for other options)</string>
     <string name="uploadVideo">Upload\nVideo</string>
     <string name="compressVideo">Compress\nVideo</string>
